Make SmartSystem a library to import
- Make class and methods
- Make logic simple and compact
- Use no logging and error handling
- Remove main light init code
- Remove extra whitespace and comments
